Add ir package

+++ src/core/range.go
@@ -1,8 +1,8 @@
 package core
 
 type Pos struct {
-	Line   uint16
-	Column uint16
+	Line   uint32
+	Column uint32
 }
 
 type Range struct {
